# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "MinIO-base data storage",
        "title": "CloudStorage",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/client/api/v1",
    "paths": {
        "/client/api/v1/delete-file": {
            "delete": {
                "description": "Delete file by user apikey and filename",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Delete a file by api",
                "parameters": [
                    {
                        "type": "string",
                        "example": "60601fee-2bf1-4721-ae6f-7636e79a0cba",
                        "description": "APIKEY (UUID)",
                        "name": "api",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "alohadance.png",
                        "description": "File name",
                        "name": "filename",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.FileResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/client/api/v1/get-file": {
            "get": {
                "description": "Get file by user apikey and filename",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Get a file by api",
                "parameters": [
                    {
                        "type": "string",
                        "example": "60601fee-2bf1-4721-ae6f-7636e79a0cba",
                        "description": "APIKEY (UUID)",
                        "name": "apikey",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "alohadance.png",
                        "description": "File name",
                        "name": "filename",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.FileInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/client/api/v1/get-files-list": {
            "get": {
                "description": "Get user files by his apikey",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Get user file list by api",
                "parameters": [
                    {
                        "type": "string",
                        "example": "60601fee-2bf1-4721-ae6f-7636e79a0cba",
                        "description": "APIKEY (UUID)",
                        "name": "api",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.FileWebResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/client/api/v1/storage/": {
            "get": {
                "description": "Page with user files",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Storage page",
                "responses": {
                    "200": {
                        "description": "Storage page",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "302": {
                        "description": "Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/client/api/v1/upload-files": {
            "post": {
                "description": "Upload file by user apikey and files from query body",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Upload a file",
                "parameters": [
                    {
                        "type": "string",
                        "example": "60601fee-2bf1-4721-ae6f-7636e79a0cba",
                        "description": "APIKEY (UUID)",
                        "name": "api",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "File to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.FileResponse"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Check if API is running",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.HealthResponse"
                        }
                    }
                }
            }
        },
        "/index": {
            "get": {
                "description": "Page to enter user apikey to get access for his storage",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Page to login",
                "responses": {
                    "200": {
                        "description": "Login page",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "302": {
                        "description": "Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.FileInfo": {
            "type": "object",
            "properties": {
                "content_type": {
                    "type": "string",
                    "example": "image/png"
                },
                "etag": {
                    "type": "string",
                    "example": "\"33a64df551425fcc55e4d42a148795d9f25f89d4\""
                },
                "last_modified": {
                    "type": "string",
                    "example": "2024-01-01T12:00:00Z"
                },
                "name": {
                    "type": "string",
                    "example": "alohadance.png"
                },
                "size": {
                    "type": "integer",
                    "example": 1024
                }
            }
        },
        "models.FileResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "new_files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.FileWebResponse"
                    }
                },
                "status": {
                    "type": "integer"
                },
                "uploaded_files": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "models.FileWebResponse": {
            "type": "object",
            "properties": {
                "create_date": {
                    "type": "string"
                },
                "file_name": {
                    "type": "string"
                },
                "file_size": {
                    "type": "string"
                },
                "file_type": {
                    "type": "string"
                }
            }
        },
        "models.HealthResponse": {
            "type": "object",
            "properties": {
                "service": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "time_stamp": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        }
    }
}